LPS-69043 Move to "commands" package
